Use actual values in step names when reporting.
When reporting failed (or succeeded) tests, the
Example values are not reported in the result. For
example: 'there are <start> cucumbers' instead of
'there are 12 cucumbers'.

This makes debugging harder (imagine you are using
several user accounts and one account fails).

This change adds/updates a 'reporting_name'
attribute to the step which will be used as the
name when serializing.

Some considerations:
- The same step will have different values for
reporting name during the execution of the tests
since a scenario outline will have different
values for each example.
- The 'name' attribute is not overwritten since it
might create a backwards incompatible change.
- Another reason why it is not overwritten is
because we need the original to format the
subsequent runs (in case of scenario outlines)