Fill in step names in cucumber report.
According to [1], when the outline scenarios are
reported in 'expand' mode, the step name
parameters are filled in.

This is a different mode than without 'expand'
mode[2], where the step names are not filled in,
but where the "examples" are reported, but not
the results of each step.

Since pytest-bdd does report the results of each
step of an outline scenario, it looks more like
the 'expand' mode, so I've added the filling
in of the parameters to make it more compliant.


[1] https://relishapp.com/cucumber/cucumber/docs/formatters/json-output-formatter#scenario-outline-expanded
[2] https://relishapp.com/cucumber/cucumber/docs/formatters/json-output-formatter#scenario-outline