Fix issue: https://github.com/pytest-dev/pytest-bdd/issues/390
Changed validation of a scenario against its examples table so
that the list of parameters defined for the scenario does not
have to be the same as the list of parameters defined in the
examples table, but can be a subset.

This allows columns to be specified in the examples table that
are not used in the scenario, but are there for future use or
purely for documentation purposes.