Don't make steps fixtures
This is an attempt at removing how steps are registered via fixtures,
and instead use a custom registry. It seems that using fixtures for that
is a bit awkward, as we can't differentiate steps from regular fixtures
apart from the parser attribute, and we need to access private pytest
attributes to do so. Given that pytest_bdd both produces and consumes
the steps, it would seem that we can remove it. As an added bonus it
should have a minor performance increase.

This is mostly a proof of concept, as I wanted to start the discussion.
Let me know if it doesn't make sense, thanks.