Fix tests/feature/test_cucumber_json.py::test_step_trace
This test was failing when run as part of the entire test suite and
passing when run on its own. When the first part of the  expected[0].id
path was changed from test_step_trace0 to test_step_trace1, the results
reversed - it passed as part of the entire suite and failed on its own.
This behavior was not observed prior to the async changed.

I suspect that this is caused by the addition of
test_async_steps.py::test_step_trace, which caused this to no longer
be the first test with the name `test_step_trace` to run as part of the
entire suite. Since there are several other tests with the same name,
it seems like the best course here would be to ensure that this test
has a unique name so that it is more resilient to ordering.